add types, remove hardcoded address, cleanup
