+++ packages/wallet/src/data/index.ts
@@ -1,3 +1,9 @@
 export { type ApiRouter, apiRouter } from './api'
-export type { ApiInput, ApiOutput, Collectible, NetworkType } from './api/types'
+export type {
+  Activity,
+  ApiInput,
+  ApiOutput,
+  Collectible,
+  NetworkType,
+} from './api/types'
 export { createAPI } from './trpc/api'
